requirements.txt
Django==4.2.16
mysqlclient==2.2.4
python-dotenv==1.0.1
djangorestframework==3.15.2
factory-boy==3.3.1
djangorestframework-simplejwt==5.3.1
django-filter==24.3